Add check for deleting single address and note

+++ src/CustomerLibraryAPI.Data/EFRepositories/AddressRepository.cs
@@ -75,6 +75,13 @@ namespace CustomerLibraryAPI.Data.EFRepositories
                 throw new NotFoundException("Address is not found.");
             }
 
+            var count = CountByCustomerId(address.CustomerId);
+
+            if (count < 2)
+            {
+                throw new NotDeletedException("Cannot delete the only address.");
+            }
+
             _context.Addresses.Remove(address);
 
             _context.SaveChanges();
